Create cache module to provide id or name get. e.g.. application id, instance id, service id…
#423
