+++ apm-collector/apm-collector-agentregister/src/main/java/org/skywalking/apm/collector/agentregister/worker/instance/dao/IInstanceDAO.java
@@ -33,6 +33,4 @@ public interface IInstanceDAO {
     void save(InstanceDataDefine.Instance instance);
 
     void updateHeartbeatTime(int instanceId, long heartbeatTime);
-
-    int getApplicationId(int applicationInstanceId);
 }
