+++ apm-collector/apm-collector-agentregister/src/main/java/org/skywalking/apm/collector/agentregister/worker/instance/dao/InstanceH2DAO.java
@@ -25,6 +25,7 @@ import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
  * @author pengys5
  */
 public class InstanceH2DAO extends H2DAO implements IInstanceDAO {
+
     @Override public int getInstanceId(int applicationId, String agentUUID) {
         return 0;
     }
@@ -43,9 +44,5 @@ public class InstanceH2DAO extends H2DAO implements IInstanceDAO {
 
     @Override public void updateHeartbeatTime(int instanceId, long heartbeatTime) {
 
-    }
-
-    @Override public int getApplicationId(int applicationInstanceId) {
-        return 0;
     }
 }
