+++ apm-collector/apm-collector-agentregister/src/main/java/org/skywalking/apm/collector/agentregister/worker/servicename/dao/ServiceNameH2DAO.java
@@ -26,20 +26,12 @@ import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
  */
 public class ServiceNameH2DAO extends H2DAO implements IServiceNameDAO {
 
-    @Override public int getServiceId(int applicationId, String serviceName) {
-        return 0;
-    }
-
     @Override public int getMaxServiceId() {
         return 0;
     }
 
     @Override public int getMinServiceId() {
         return 0;
-    }
-
-    @Override public String getServiceName(int serviceId) {
-        return null;
     }
 
     @Override public void save(ServiceNameDataDefine.ServiceName serviceName) {
