+++ apm-collector/apm-collector-agentstream/pom.xml
@@ -61,5 +61,10 @@
             <artifactId>apm-collector-agentregister</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.skywalking</groupId>
+            <artifactId>apm-collector-cache</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 </project>
