+++ apm-collector/apm-collector-cache/src/main/java/org/skywalking/apm/collector/cache/ServiceIdCache.java
@@ -16,40 +16,40 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.collector.ui.cache;
+package org.skywalking.apm.collector.cache;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
+import org.skywalking.apm.collector.cache.dao.IServiceNameDAO;
 import org.skywalking.apm.collector.core.util.Const;
 import org.skywalking.apm.collector.storage.dao.DAOContainer;
-import org.skywalking.apm.collector.ui.dao.IServiceNameDAO;
 
 /**
  * @author pengys5
  */
-public class ServiceNameCache {
+public class ServiceIdCache {
 
     //TODO size configuration
-    private static Cache<Integer, String> CACHE = CacheBuilder.newBuilder().maximumSize(10000).build();
+    private static Cache<String, Integer> CACHE = CacheBuilder.newBuilder().maximumSize(1000).build();
 
-    public static String get(int serviceId) {
+    public static int get(int applicationId, String serviceName) {
         try {
-            return CACHE.get(serviceId, () -> {
+            return CACHE.get(applicationId + Const.ID_SPLIT + serviceName, () -> {
                 IServiceNameDAO dao = (IServiceNameDAO)DAOContainer.INSTANCE.get(IServiceNameDAO.class.getName());
-                return dao.getServiceName(serviceId);
+                return dao.getServiceId(applicationId, serviceName);
             });
         } catch (Throwable e) {
-            return Const.EXCEPTION;
+            return 0;
         }
     }
 
-    public static String getForUI(int serviceId) {
-        String serviceName = get(serviceId);
-        if (serviceName.equals("Unknown")) {
+    public static int getForUI(int applicationId, String serviceName) {
+        int serviceId = get(applicationId, serviceName);
+        if (serviceId == 0) {
             IServiceNameDAO dao = (IServiceNameDAO)DAOContainer.INSTANCE.get(IServiceNameDAO.class.getName());
-            serviceName = dao.getServiceName(serviceId);
-            CACHE.put(serviceId, serviceName);
+            serviceId = dao.getServiceId(applicationId, serviceName);
+            CACHE.put(applicationId + Const.ID_SPLIT + serviceName, serviceId);
         }
-        return serviceName;
+        return serviceId;
     }
 }
