+++ apm-collector/apm-collector-cache/src/main/java/org/skywalking/apm/collector/cache/ServiceNameCache.java
@@ -16,40 +16,40 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.collector.ui.cache;
+package org.skywalking.apm.collector.cache;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
+import org.skywalking.apm.collector.cache.dao.IServiceNameDAO;
 import org.skywalking.apm.collector.core.util.Const;
 import org.skywalking.apm.collector.storage.dao.DAOContainer;
-import org.skywalking.apm.collector.ui.dao.IServiceNameDAO;
 
 /**
  * @author pengys5
  */
-public class ServiceIdCache {
+public class ServiceNameCache {
 
     //TODO size configuration
-    private static Cache<String, Integer> CACHE = CacheBuilder.newBuilder().maximumSize(1000).build();
+    private static Cache<Integer, String> CACHE = CacheBuilder.newBuilder().maximumSize(10000).build();
 
-    public static int get(int applicationId, String serviceName) {
+    public static String get(int serviceId) {
         try {
-            return CACHE.get(applicationId + Const.ID_SPLIT + serviceName, () -> {
+            return CACHE.get(serviceId, () -> {
                 IServiceNameDAO dao = (IServiceNameDAO)DAOContainer.INSTANCE.get(IServiceNameDAO.class.getName());
-                return dao.getServiceId(applicationId, serviceName);
+                return dao.getServiceName(serviceId);
             });
         } catch (Throwable e) {
-            return 0;
+            return Const.EXCEPTION;
         }
     }
 
-    public static int getForUI(int applicationId, String serviceName) {
-        int serviceId = get(applicationId, serviceName);
-        if (serviceId == 0) {
+    public static String getForUI(int serviceId) {
+        String serviceName = get(serviceId);
+        if (serviceName.equals("Unknown")) {
             IServiceNameDAO dao = (IServiceNameDAO)DAOContainer.INSTANCE.get(IServiceNameDAO.class.getName());
-            serviceId = dao.getServiceId(applicationId, serviceName);
-            CACHE.put(applicationId + Const.ID_SPLIT + serviceName, serviceId);
+            serviceName = dao.getServiceName(serviceId);
+            CACHE.put(serviceId, serviceName);
         }
-        return serviceId;
+        return serviceName;
     }
 }
