+++ apm-collector/apm-collector-cache/src/main/java/org/skywalking/apm/collector/cache/dao/ApplicationH2DAO.java
@@ -16,9 +16,9 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.collector.agentregister.worker.application.dao;
+package org.skywalking.apm.collector.cache.dao;
 
-import org.skywalking.apm.collector.storage.define.register.ApplicationDataDefine;
+import org.skywalking.apm.collector.client.h2.H2Client;
 import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
 
 /**
@@ -26,15 +26,12 @@ import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
  */
 public class ApplicationH2DAO extends H2DAO implements IApplicationDAO {
 
-    @Override public int getMaxApplicationId() {
-        return 0;
+    @Override public int getApplicationId(String applicationCode) {
+        H2Client client = getClient();
+        return 100;
     }
 
-    @Override public int getMinApplicationId() {
-        return 0;
-    }
-
-    @Override public void save(ApplicationDataDefine.Application application) {
-
+    @Override public String getApplicationCode(int applicationId) {
+        return null;
     }
 }
