+++ apm-collector/apm-collector-cache/src/main/java/org/skywalking/apm/collector/cache/dao/InstanceH2DAO.java
@@ -16,20 +16,15 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.collector.ui.dao;
+package org.skywalking.apm.collector.cache.dao;
 
 import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
 
 /**
  * @author pengys5
  */
-public class ServiceNameH2DAO extends H2DAO implements IServiceNameDAO {
-
-    @Override public String getServiceName(int serviceId) {
-        return null;
-    }
-
-    @Override public int getServiceId(int applicationId, String serviceName) {
+public class InstanceH2DAO extends H2DAO implements IInstanceDAO {
+    @Override public int getApplicationId(int applicationInstanceId) {
         return 0;
     }
 }
