+++ apm-collector/apm-collector-cache/src/main/java/org/skywalking/apm/collector/cache/dao/ServiceNameH2DAO.java
@@ -16,13 +16,20 @@
  * Project repository: https://github.com/OpenSkywalking/skywalking
  */
 
-package org.skywalking.apm.collector.ui.dao;
+package org.skywalking.apm.collector.cache.dao;
+
+import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
 
 /**
  * @author pengys5
  */
-public interface IServiceNameDAO {
-    String getServiceName(int serviceId);
+public class ServiceNameH2DAO extends H2DAO implements IServiceNameDAO {
 
-    int getServiceId(int applicationId, String serviceName);
+    @Override public String getServiceName(int serviceId) {
+        return null;
+    }
+
+    @Override public int getServiceId(int applicationId, String serviceName) {
+        return 0;
+    }
 }
