+++ apm-collector/apm-collector-cache/src/main/resources/META-INF/defines/es_dao.define
@@ -1,0 +1,3 @@
+org.skywalking.apm.collector.cache.dao.ApplicationEsDAO
+org.skywalking.apm.collector.cache.dao.InstanceEsDAO
+org.skywalking.apm.collector.cache.dao.ServiceNameEsDAO
