+++ apm-collector/apm-collector-cache/src/main/resources/META-INF/defines/h2_dao.define
@@ -1,0 +1,3 @@
+org.skywalking.apm.collector.cache.dao.ApplicationH2DAO
+org.skywalking.apm.collector.cache.dao.InstanceH2DAO
+org.skywalking.apm.collector.cache.dao.ServiceNameH2DAO
