+++ apm-collector/apm-collector-ui/pom.xml
@@ -51,5 +51,10 @@
             <artifactId>apm-collector-storage</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.skywalking</groupId>
+            <artifactId>apm-collector-cache</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 </project>
