+++ apm-collector/apm-collector-ui/src/main/resources/META-INF/defines/es_dao.define
@@ -4,8 +4,6 @@ org.skywalking.apm.collector.ui.dao.NodeReferenceEsDAO
 org.skywalking.apm.collector.ui.dao.SegmentCostEsDAO
 org.skywalking.apm.collector.ui.dao.GlobalTraceEsDAO
 org.skywalking.apm.collector.ui.dao.SegmentEsDAO
-org.skywalking.apm.collector.ui.dao.ApplicationEsDAO
-org.skywalking.apm.collector.ui.dao.ServiceNameEsDAO
 org.skywalking.apm.collector.ui.dao.InstanceEsDAO
 org.skywalking.apm.collector.ui.dao.InstPerformanceEsDAO
 org.skywalking.apm.collector.ui.dao.CpuMetricEsDAO
