+++ apm-collector/apm-collector-ui/src/main/resources/META-INF/defines/h2_dao.define
@@ -4,8 +4,6 @@ org.skywalking.apm.collector.ui.dao.NodeReferenceH2DAO
 org.skywalking.apm.collector.ui.dao.SegmentCostH2DAO
 org.skywalking.apm.collector.ui.dao.GlobalTraceH2DAO
 org.skywalking.apm.collector.ui.dao.SegmentH2DAO
-org.skywalking.apm.collector.ui.dao.ApplicationH2DAO
-org.skywalking.apm.collector.ui.dao.ServiceNameH2DAO
 org.skywalking.apm.collector.ui.dao.InstanceH2DAO
 org.skywalking.apm.collector.ui.dao.InstPerformanceH2DAO
 org.skywalking.apm.collector.ui.dao.GCMetricH2DAO
